+++ NPCCreater/bName.txt
@@ -1,0 +1,256 @@
+Abel
+Abelard
+Abraham
+Addison
+Alaire
+Albin
+Aldebrand
+Aldous
+Aleyn
+Alistair
+Ancelot
+Anselm
+Aran
+Arnald
+Arnott
+Arthur
+Augustine
+Aylmer
+Baderon
+Baldric
+Bardolf
+Bartholomew
+Bayard
+Belmont
+Benedict
+Beneger
+Bernard
+Berndan
+Bertram
+Bertrand
+Blackburn
+Blavier
+Bouchard
+Boyle
+Bran
+Brice
+Brien
+Bruce
+Bryce
+Cain
+Cameron
+Caplan
+Carmine
+Caspar
+Ceadda
+Chamberlain
+Charlys
+Chartain
+Claudien
+Clifton
+Clive
+Cole
+Colson
+Conphas
+Cornell
+Coster
+Cutbert
+Cuthbert
+Cyriac
+Daimbert
+Dalmas
+Danyell
+Dauid
+Davyd
+Dawson
+Deitrich
+Denston
+Derwin
+Deryk
+Donner
+Drake
+Drew
+Drystan
+Eadbert
+Ealdwine
+Edmund
+Edwyn
+Eldred
+Eleazar
+Emanuel
+Emerick
+Erasmus
+Erik
+Esmond
+Esmour
+Esperaunce
+Etgar
+Ethelbert
+Ethelred
+Eustace
+Fawkes
+Fiebras
+Flambard
+Folke
+Foxe
+Francis
+Frederick
+Frederyk
+Fulke
+Galfrid
+Ganelon
+Gared
+Gauwyn
+Gembert
+Geoffrey
+Gerald
+Gerbold
+Gerhardt
+Gerland
+Goddard
+Godebert
+Godfrey
+Gregory
+Grimbald
+Gryffen
+Guston
+Gwayne
+Gylbart
+Gyles
+Habreham
+Hadrian
+Haimirich
+Halstein
+Hamon
+Heinlein
+Hewrey
+Humphrey
+Ingham
+Ingram
+Isleton
+Ivan
+Jakys
+Jeger
+Jenlyns
+Johannes
+Jonathas
+Joseph
+Josias
+Joyce
+Kain
+Kennard
+Kenrick
+Kerrich
+Khellus
+Kimball
+Kinnison
+Kurtz
+Ladislas
+Lambert
+Lars
+Laurence
+Laurentius
+Leavold
+Lefwyne
+Lennard
+Leopold
+Littlejohn
+Lloyd
+Lodwicke
+Lowell
+Madison
+Mainfroi
+Mansel
+Mathye
+Morgant
+Morys
+Myles
+Nathaniel
+Navarre
+Neale
+Noes
+Norman
+Olyver
+Orrick
+Orwen
+Osric
+Oswyn
+Owyne
+Parnell
+Patrick
+Paul
+Percival
+Peter
+Philippe
+Piers
+Powle
+Radcliffe
+Radolf
+Raffe
+Randall
+Randwulf
+Rauffe
+Raulin
+Redwald
+Reeve
+Reginald
+Reinholdt
+Reynard
+Reyner
+Reynfred
+Ricard
+Richarde
+Rickeman
+Ridel
+Robert
+Robyn
+Roger
+Rolfe
+Ronald
+Roundelph
+Rowland
+Samson
+Sandre
+Sevrin
+Sighard
+Sigurdh
+Simond
+Singleton
+Solyeuse
+Spenser
+Stewart
+Swift
+Symon
+Symond
+Taran
+Taylor
+Templeton
+Theodore
+Thomas
+Thrydwulf
+Timothy
+Tristan
+Turstin
+Ulric
+Valentyne
+Vannes
+Victor
+Voyce
+Vyncent
+Wadard
+Walter
+Warin
+Wauter
+Werner
+Wilfrid
+Wilham
+Willielmus
+Wineburg
+Wolfstan
+Wymon
+Wymond
+Wystan
+Ywain
+Zacheus
+Zell
+Zerig
