+++ NPCCreater/class.txt
@@ -1,0 +1,6 @@
+Fighter 
+Archer 
+Healer
+Sorcerer 
+Bard 
+Librarian
