+++ NPCCreater/gName.txt
@@ -1,0 +1,256 @@
+Adela
+Adelaide
+Admiranda
+Aeditha
+Aelina
+Agnys
+Alainne
+Alianore
+Alison
+Alyne
+Alys
+Amelia
+Amice
+Amphelice
+Angelet
+Anna
+Annabel
+Anne
+Anthoinette
+Anys
+Arabella
+Arlette
+Atilda
+Aubrey
+Audrye
+Ava
+Avelin
+Avelyn
+Averil
+Ayleth
+Baterich
+Bathsua
+Beatrix
+Bellinda
+Bertana
+Berte
+Bess
+Brangwine
+Braya
+Brunhild
+Bryde
+Caesaria
+Carmen
+Casandra
+Cecilia
+Cecily
+Celeste
+Celestine
+Celestria
+Cenota
+Chloe
+Christabel
+Cicely
+Clarimond
+Claudia
+Clemence
+Collys
+Concessa
+Constance
+Cornelia
+Crestian
+Cristiana
+Cwengyth
+Cyndra
+Cynewyn
+Damaris
+Dametta
+Decima
+Deloys
+Denys
+Diamanda
+Dionisia
+Dominy
+Dorcas
+Dorothe
+Durilda
+Dyana
+Edelinne
+Edithe
+Eilonwy
+Elaisse
+Ele
+Eleanor
+Elewys
+Ellerete
+Ellie
+Elsebee
+Elyn
+Elynor
+Elyzabeth
+Emblyn
+Emeline
+Emeny
+Emeria
+Emery
+Emilie
+Emlinie
+Emmet
+Eschina
+Eschiva
+Esdeline
+Esmenet
+Estienne
+Estrild
+Ethelia
+Eugenia
+Eustacia
+Eva
+Evelyn
+Felice
+Florens
+Frances
+Francisca
+Frideswide
+Fridgia
+Gaynor
+Germainne
+Gethrude
+Gillian
+Giselle
+Glenda
+Gloriana
+Guinevere
+Gylda
+Helena
+Helenor
+Helvynya
+Hester
+Hilda
+Hildegard
+Hilith
+Imedia
+Isabella
+Isemeine
+Ismay
+Ismenia
+Isolde
+Jaane
+Jacquette
+Jeanne
+Jellion
+Jemime
+Jenet
+Jenyfer
+Jessamine
+Jillian
+Jocea
+Jocelyn
+Joleicia
+Jolline
+Josephine
+Josian
+Josiane
+Joyse
+Judithe
+Judye
+Juliana
+Julyan
+June
+Justina
+Katelyn
+Kath
+Katherine
+Katrine
+Kinborow
+Latisha
+Lauda
+Leofwen
+Leofwynn
+Letita
+Lettice
+Linette
+Linyeve
+Lora
+Maddeline
+Maerwynn
+Maisenta
+Malin
+Margarete
+Margeria
+Margry
+Maria
+Maronne
+Marsilia
+Martine
+Mathild
+Melodie
+Melusine
+Meredithe
+Merewyn
+Merilda
+Meryell
+Millicent
+Minerva
+Mirabelle
+Morgayne
+Muriel
+Murienne
+Mydrede
+Nesta
+Nicholina
+Nicia
+Nicolaa
+Olyffe
+Ophellia
+Ottilia
+Paige
+Parnella
+Pelinne
+Penelope
+Petronilla
+Placencia
+Prudence
+Pulmia
+Purnell
+Rebeccah
+Rianna
+Richenda
+Rosa
+Rosalind
+Rosamund
+Rose
+Roysia
+Rychyld
+Samantha
+Sanche
+Sarra
+Scarlet
+Selphina
+Sence
+Serendipity
+Somerhild
+Sreda
+Sybell
+Sylphie
+Syndony
+Sysley
+Systeleley
+Tansa
+Temperance
+Theda
+Theresa
+Thomasine
+Thomasyn
+Thora
+Tiphina
+Tristana
+Ursula
+Vrsela
+Wenyld
+Willmott
+Wulfhilda
+Wynefreede
+Yedythe
+Ysabel
+Ysmeina
