+++ README.md
@@ -3,11 +3,11 @@
 PROJECT-0 (NPCGEN) is a CLI application which will assist a player in creating a non-player role-playing characters for use in table top RPGs. Currently following D20PFSRD rules found at https://www.d20pfsrd.com/classes/npc-classes/Non-player-characters/
 
 ## Plans
-
+- [ ] Male or Female Character
 - [ ] Get NPC name from user and display back
     [x] Get NPC name during runtime
     [ ] Get NPC name from a flag
-    [ ] Get a random NPC name
+    [ ] Get a random NPC name (names stored in a txt file, names from FINAL FANTASY TATICS found at http://dustin.wikidot.com/names)
 - [ ] Get NPC class from user and compare against allowed classes, return approved class
     [x] Get NPC class during runtime
     [ ] Get NPC class from a flag
